switching to github actions
This commit removes Buildkite and uses the most excellent GitHub actions instead.
