+++ .vscode/spellright.dict
@@ -1,7 +1,6 @@
 Alignas
 Alignof
 Armin
-Buildkite
 Doctest
 Kirienko
 Lenfesty
