+++ CONTRIBUTING.rst
@@ -82,12 +82,12 @@ To run the language verification build you'll need to use a different docker con
 
     docker pull uavcan/c_cpp:ubuntu-18.04
     docker run --rm -it -v $PWD:/repo uavcan/c_cpp:ubuntu-18.04
-    ./.buildkite/verify.py -l c
-    ./.buildkite/verify.py -l cpp
+    ./.github/verify.py -l c
+    ./.github/verify.py -l cpp
 
 The verify.py script is a simple commandline generator for our cmake scripts. Use help for details::
 
-    ./.buildkite/verify.py --help
+    ./.github/verify.py --help
 
 Files Generated by the Tests
 ================================================
@@ -227,42 +227,6 @@ At the end of the mypy run we generate the following summaries:
 - .tox/mypy/tmp/mypy-report-script/index.txt
 
 ************************************************
-Buildkite
-************************************************
-
-If you have Docker, the `Buildkite agent`_, and the Buildkite CLI installed you'll be able to use the `Buildkite cli`_ to
-run local builds. For example::
-
-    bk local run
-
-The run command will execute all build step in series so it could be really slooooowwwww. Use filters to build
-select individual steps to speed things up::
-
-    bk local run --filter="verification build"
-
-Make sure you are in the nunavut root directory when you run these commands. Use the :code:`--debug` flag if you run
-into problems::
-
-    bk --debug local run
-
-Administrating Buildkite on AWS
-================================================
-
-The PyPI upload keys should be rotated periodically. To do this you'll need to be an administrator of
-our Buildkite `AWS CloudFormation`_ stack and of our PyPI UAVCAN organization.
-
-    1. Download the buildkite-managedsecretsbucket-xxxxxxxx/nunavut-release/env s3 artifact.
-    2. In your PyPI account settings create a new API key scoped only to the nunavut project and replace
-       the one in the downloaded env file.
-    3. Upload the modified env file::
-
-        aws s3 cp --acl private --sse aws:kms ~/Downloads/env "s3://buildkite-managedsecretsbucket-xxxxxxxx/nunavut-release/env"
-
-    4. Back in the PyPI keys list, delete any keys that are older than the one previously in use. You can keep the key
-       you just rotated until you rotate the new key.
-
-
-************************************************
 Nunavut Verification Suite
 ************************************************
 
@@ -369,7 +333,4 @@ json and we can get rid of this ugly hack. Sorry.
 .. _`UAVCAN website`: http://uavcan.org
 .. _`UAVCAN forum`: https://forum.uavcan.org
 .. _`nunavut on read the docs`: https://nunavut.readthedocs.io/en/latest/index.html
-.. _`AWS CloudFormation`: https://aws.amazon.com/cloudformation/
 .. _`VSCode Remote Containers`: https://code.visualstudio.com/docs/remote/containers
-.. _`Buildkite cli`: https://github.com/buildkite/cli
-.. _`Buildkite agent`: https://github.com/buildkite/agent
