+++ sonar-project.properties
@@ -27,5 +27,5 @@ sonar.sourceEncoding=UTF-8
 
 sonar.python.coverage.reportPaths=.tox/report/tmp/coverage.xml
 
-# Python3.7 is our "target" version so we use that test run's report.
-sonar.python.xunit.reportPat=.tox/py37-test/tmp/xunit-result.xml
+# Python3.9 is our "target" version so we use that test run's report.
+sonar.python.xunit.reportPath=.tox/py39-test/tmp/xunit-result.xml
