+++ src/nunavut/version.py
@@ -7,7 +7,7 @@
 .. autodata:: __version__
 """
 
-__version__ = "1.6.0"  #: The version number used in the release of nunavut to pypi.
+__version__ = "1.6.1"  #: The version number used in the release of nunavut to pypi.
 
 
 __license__ = "MIT"
