+++ tox.ini
@@ -5,7 +5,7 @@
 # to 3.9 since typing is still a bit fluid accross versions.
 #
 [tox]
-envlist = {py35,py36,py37,py38,py39}-{test,nnvg,doctest,rstdoctest},lint,mypy,report,docs
+envlist = {py35,py36,py37,py38,py39}-{test,nnvg,doctest,rstdoctest},lint,report,docs
 
 
 [base]
@@ -126,12 +126,8 @@ setenv =
     PYTHONPATH={toxinidir}/test
 
 passenv =
-    BUILDKITE
-    BUILDKITE_*
-    PYPI_PASSWORD
-    SONARQUBE_TOKEN
-    NUNAVUT_MAJOR_MINOR_VERSION
-    NUNAVUT_FULL_VERSION
+    GITHUB
+    GITHUB_*
 
 deps =
     test,nnvg,doctest,rstdoctest: {[base]deps}
@@ -149,6 +145,7 @@ commands =
 
     test: coverage run \
     test:    -m pytest {posargs} --basetemp={envtmpdir} -p "no:cacheprovider" \
+    test:    --junit-xml={envtmpdir}/xunit-result.xml \
     test:    --rootdir={toxinidir} \
     test:    {toxinidir}/test
 
@@ -198,38 +195,31 @@ commands =
     coverage xml -o {envtmpdir}/coverage.xml
 
 
-[testenv:mypy]
-deps =
-    mypy
-    lxml
-    types-pkg_resources
-    types-PyYAML
-    pydsdl
-
-basepython = python3.9
-commands =
-    mypy -m nunavut \
-         -m nunavut.jinja \
-         -p nunavut.lang \
-        --cache-dir {envtmpdir} \
-        --txt-report {envtmpdir}/mypy-report-lib \
-        --config-file {toxinidir}/tox.ini
-
-
 [testenv:lint]
-skip_install = true
+basepython = python3.9
 deps =
     black
     flake8
     doc8
     rstcheck
     Pygments
+    mypy
+    lxml
+    types-pkg_resources
+    types-PyYAML
+    pydsdl
 
 commands =
     flake8 --benchmark --tee --output-file={envtmpdir}/flake8.txt --filename=*.py --exclude=**/jinja2/*,**/markupsafe/* src
     black --check --line-length 120 --force-exclude '(/jinja2/|/markupsafe\/)' src
     doc8 {toxinidir}/docs
     rstcheck --ignore-directives D001 --ignore-directives autofunction,automodule,argparse --ignore-roles mod,func,class -r {toxinidir}
+    mypy -m nunavut \
+         -m nunavut.jinja \
+         -p nunavut.lang \
+        --cache-dir {envtmpdir} \
+        --txt-report {envtmpdir}/mypy-report-lib \
+        --config-file {toxinidir}/tox.ini
 
 
 [testenv:package]
@@ -243,29 +233,8 @@ commands =
         bdist_wheel -d {toxworkdir}/package/dist \
                     -b {envtmpdir}/tmp \
                     -k \
-                    --build-number {env:BUILDKITE_BUILD_NUMBER:0}
+                    --build-number {env:GITHUB_RUN_ID:0}
     twine check {toxworkdir}/package/dist/*
-
-
-[testenv:sonar]
-whitelist_externals = sonar-scanner
-
-commands=
-    sonar-scanner \
-        -Dsonar.login={env:SONARQUBE_TOKEN:} \
-        -Dsonar.buildString={env:BUILDKITE_BUILD_NUMBER:0} \
-        -Dsonar.projectVersion={env:NUNAVUT_MAJOR_MINOR_VERSION:0} \
-        -Dsonar.working.directory={envtmpdir}/.scannerwork
-
-
-[testenv:upload]
-skip_install = true
-deps =
-    twine
-
-commands =
-    twine upload -u __token__ -p {env:PYPI_PASSWORD:} {toxworkdir}/package/dist/*
-
 
 [testenv:local]
 usedevelop = true
@@ -273,5 +242,4 @@ deps =
     {[base]deps}
     {[dev]deps}
     {[testenv:docs]deps}
-    {[testenv:mypy]deps}
     {[testenv:lint]deps}
